Move map constants, update helm-docs
